投稿機能の完成

+++ app/controllers/posts_controller.rb
@@ -1,10 +1,11 @@
 class PostsController < ApplicationController
   def index
-    @posts = current_user.posts
+    @posts = Post.all
   end
 
   def show
-    @post = current_user.posts.find(params[:id])
+    @post = Post.find(params[:id])
+    #@post = current_user.posts.find(params[:id])以前かいたやつ
   end
 
   def new
@@ -12,7 +13,18 @@ class PostsController < ApplicationController
   end
 
   def edit
-    @posts = current_user.posts.find(params[:id])
+    if current_user.admin?
+      @posts = Post.find(params[:id])
+    else
+      redirect_to root_path, notice: '権限がありません'
+    end
+    #if current_user.admin? then
+    #  @posts = Post.find(params[:id])
+    #elsif current_user && current_user.posts.find_by(params[:id]).empty? then
+    #  @posts = current_user.posts.find(params[:id])
+    #else
+    #  redirect_to root_path, notice: '権限がありません'
+    #end
   end
 
   def update
@@ -26,9 +38,13 @@ class PostsController < ApplicationController
   end
 
   def destroy
-    post = current_user.posts.find(params[:id])
-    post.destroy
-    redirect_to root_path, notice: 'Success!'
+    if current_user.admin?
+      post = Post.find(params[:id])
+      post.destroy
+      redirect_to root_path, notice: 'Success!'
+    else
+      redirect_to root_path, notice: '権限がありません'
+    end
   end
 
   def create
